requirements.txt
PyYAML==3.12
Flask==0.11.1
marshmallow==2.9.1

pytest==3.5.1
pytest-cov==2.5.1